+++ example/lib/vector_map_tiles_pmtiles/page.dart
@@ -1,18 +1,31 @@
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_map/flutter_map.dart';
+import 'package:flutter_map_plugins_example/common/attribution_widget.dart';
 import 'package:latlong2/latlong.dart';
 import 'package:vector_map_tiles/vector_map_tiles.dart';
 import 'package:vector_map_tiles_pmtiles/vector_map_tiles_pmtiles.dart';
-import 'package:vector_tile_renderer/vector_tile_renderer.dart';
+import 'package:vector_tile_renderer/vector_tile_renderer.dart' as vtr;
 
-// It could be that the hosted PMTiles file is no longer available.
-// Check https://maps.protomaps.com/builds/ to get an up to date build.
-// TODO: use your own tile source https://docs.protomaps.com/pmtiles/cloud-storage
-const tileSource = 'https://build.protomaps.com/20240128.pmtiles';
+/// TODO: use your own tile source https://docs.protomaps.com/pmtiles/cloud-storage
+/// This can be a hosted file or local file in your file system,
+/// However, flutter assets are not supported.
+const String tileSource =
+    'https://raw.githubusercontent.com/protomaps/PMTiles/main/spec/v3/protomaps(vector)ODbL_firenze.pmtiles';
 
 class VectorMapTilesPmTilesPage extends StatelessWidget {
-  final Future<PmTilesVectorTileProvider> _futureTileProvider =
-      PmTilesVectorTileProvider.fromSource(tileSource);
+  /// The theme specifies the look of the rendered map.
+  ///
+  /// Note: Styles from Mapbox, OpenMapTiles and others and not compatible
+  /// with Protomaps styles.
+  final vtr.Theme mapTheme = ProtomapsThemes.light(
+    logger: kDebugMode ? const vtr.Logger.console() : null,
+  );
+
+  final _futureTileProvider = PmTilesVectorTileProvider.fromSource(
+    tileSource,
+    silenceTileNotFound: true,
+  );
 
   VectorMapTilesPmTilesPage({super.key});
 
@@ -30,17 +43,20 @@ class VectorMapTilesPmTilesPage extends StatelessWidget {
             final tileProvider = snapshot.data!;
             return FlutterMap(
               options: const MapOptions(
-                initialZoom: 1,
-                initialCenter: LatLng(0, 0),
-                maxZoom: 3.49,
+                initialCenter: LatLng(43.787942, 11.237517), // firenze
+                maxZoom: 18,
+                minZoom: 0,
               ),
               children: [
                 VectorTileLayer(
-                  theme: ProvidedThemes.lightTheme(),
+                  // disable the file cache when you change the PMTiles source
+                  fileCacheTtl: Duration.zero,
+                  theme: mapTheme,
                   tileProviders: TileProviders({
-                    'openmaptiles': tileProvider,
+                    'protomaps': tileProvider,
                   }),
                 ),
+                const OsmAttributionWidget(),
               ],
             );
           }