+++ vector_map_tiles_pmtiles/CHANGELOG.md
@@ -1,3 +1,10 @@
+## [1.1.0] 2024-02-10
+
+- Add `ProtomapsThemes` with all Protomaps basemap themes.
+- Add the parameter `bool silenceTileNotFound` to silence "tile not found"
+  exceptions (`TileNotFoundException`).
+- Update example and docs
+
 ## [1.0.1] 2024-01-30
 
 - Improve documentation